+++ frontend/src/pages/Player/RightPlayerPanel/components/NetworkResourcePanel/NetworkResourcePanel.tsx
@@ -1,12 +1,9 @@
 import { PreviousNextGroup } from '@components/PreviousNextGroup/PreviousNextGroup'
-import { TableList, TableListItem } from '@components/TableList/TableList'
-import { ErrorObject } from '@graph/schemas'
 import {
 	Ariakit,
 	Badge,
 	Box,
 	ButtonIcon,
-	Callout,
 	Heading,
 	IconSolidArrowCircleRight,
 	IconSolidX,
@@ -18,43 +15,19 @@ import {
 import usePlayerConfiguration from '@pages/Player/PlayerHook/utils/usePlayerConfiguration'
 import { useReplayerContext } from '@pages/Player/ReplayerContext'
 import { useResourcesContext } from '@pages/Player/ResourcesContext/ResourcesContext'
-import RequestMetrics from '@pages/Player/Toolbar/DevToolsWindow/ResourcePage/components/RequestMetrics/RequestMetrics'
-import { UnknownRequestStatusCode } from '@pages/Player/Toolbar/DevToolsWindowV2/NetworkPage/NetworkPage'
-import {
-	formatSize,
-	getHighlightRequestId,
-	getNetworkResourcesDisplayName,
-	NetworkResource,
-} from '@pages/Player/Toolbar/DevToolsWindowV2/utils'
-import { REQUEST_INITIATOR_TYPES } from '@pages/Player/utils/utils'
-import { CodeBlock } from '@pages/Setup/CodeBlock/CodeBlock'
+import { NetworkResource } from '@pages/Player/Toolbar/DevToolsWindowV2/utils'
 import analytics from '@util/analytics'
 import { playerTimeToSessionAbsoluteTime } from '@util/session/utils'
-import { formatTime, MillisToMinutesAndSeconds } from '@util/time'
-import React, { useCallback, useEffect, useMemo, useRef, useState } from 'react'
+import { MillisToMinutesAndSeconds } from '@util/time'
+import React, { useCallback, useEffect, useMemo, useState } from 'react'
 import { useHotkeys } from 'react-hotkeys-hook'
 
-import { LinkButton } from '@/components/LinkButton'
-import LoadingBox from '@/components/LoadingBox'
-import { useGetErrorGroupsOpenSearchQuery } from '@/graph/generated/hooks'
 import { useActiveNetworkResourceId } from '@/hooks/useActiveNetworkResourceId'
-import { useProjectId } from '@/hooks/useProjectId'
-import { ErrorFeedCard } from '@/pages/ErrorsV2/ErrorFeedCard/ErrorFeedCard'
-import { FullScreenContainer } from '@/pages/LogsPage/LogsTable/FullScreenContainer'
-import { LogsTable } from '@/pages/LogsPage/LogsTable/LogsTable'
-import { SearchForm } from '@/pages/LogsPage/SearchForm/SearchForm'
-import {
-	DEFAULT_LOGS_OPERATOR,
-	stringifyLogsQuery,
-} from '@/pages/LogsPage/SearchForm/utils'
-import { useGetLogs } from '@/pages/LogsPage/useGetLogs'
-import {
-	RightPanelView,
-	usePlayerUIContext,
-} from '@/pages/Player/context/PlayerUIContext'
+import { NetworkResourceErrors } from '@/pages/Player/RightPlayerPanel/components/NetworkResourcePanel/NetworkResourceErrors'
+import { NetworkResourceInfo } from '@/pages/Player/RightPlayerPanel/components/NetworkResourcePanel/NetworkResourceInfo'
+import { NetworkResourceLogs } from '@/pages/Player/RightPlayerPanel/components/NetworkResourcePanel/NetworkResourceLogs'
 import { WebSocketMessages } from '@/pages/Player/RightPlayerPanel/components/WebSocketMessages/WebSocketMessages'
 import { useWebSocket } from '@/pages/Player/WebSocketContext/WebSocketContext'
-import { useParams } from '@/util/react-router/useParams'
 
 import * as styles from './NetworkResourcePanel.css'
 
@@ -296,7 +269,7 @@ function NetworkResourceDetails({
 				pages={{
 					[NetworkRequestTabs.Info]: {
 						page: (
-							<NetworkResourceData
+							<NetworkResourceInfo
 								selectedNetworkResource={resource}
 								networkRecordingEnabledForSession={
 									session?.enable_recording_network_contents ||
@@ -306,7 +279,12 @@ function NetworkResourceDetails({
 						),
 					},
 					[NetworkRequestTabs.Errors]: {
-						page: <ErrorsData resource={resource} hide={hide} />,
+						page: (
+							<NetworkResourceErrors
+								resource={resource}
+								hide={hide}
+							/>
+						),
 					},
 					[NetworkRequestTabs.Logs]: {
 						page: (
@@ -498,7 +476,7 @@ function WebSocketDetails({
 				pages={{
 					[WebSocketTabs.Headers]: {
 						page: (
-							<NetworkResourceData
+							<NetworkResourceInfo
 								selectedNetworkResource={resource}
 								networkRecordingEnabledForSession={
 									session?.enable_recording_network_contents ||
@@ -522,599 +500,5 @@ function WebSocketDetails({
 				pageContainerClass={styles.pageContainer}
 			/>
 		</>
-	)
-}
-
-enum NetworkResourceMeta {
-	General = 'General',
-	RequestHeaders = 'Request Headers',
-	RequestPayload = 'Request Payload',
-	ResponseHeaders = 'Response Headers',
-	ResponsePayload = 'Response Payload',
-}
-
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-function NetworkResourceData({
-	selectedNetworkResource,
-	networkRecordingEnabledForSession,
-}: {
-	selectedNetworkResource?: NetworkResource
-	networkRecordingEnabledForSession: boolean
-}) {
-	const requestHeadersData: TableListItem[] = []
-	const requestPayloadData: TableListItem[] = []
-	const responseHeadersData: TableListItem[] = []
-	const responsePayloadData: TableListItem[] = []
-
-	const generalData: TableListItem[] = [
-		{
-			keyDisplayValue: 'Request URL',
-			valueDisplayValue: selectedNetworkResource?.name ?? 'Unknown',
-		},
-		{
-			keyDisplayValue: 'Method',
-			valueDisplayValue:
-				selectedNetworkResource?.requestResponsePairs?.request.verb ||
-				'GET',
-		},
-		{
-			keyDisplayValue: 'Status',
-			valueDisplayValue:
-				selectedNetworkResource?.initiatorType === 'websocket'
-					? 101
-					: selectedNetworkResource?.requestResponsePairs?.response
-							.status ?? (
-							<UnknownRequestStatusCode
-								networkRequestAndResponseRecordingEnabled={
-									networkRecordingEnabledForSession
-								}
-							/>
-					  ),
-			valueInfoTooltipMessage:
-				selectedNetworkResource?.requestResponsePairs?.response
-					.status === 0 &&
-				'This request was blocked on the client. The usually reason is a browser extension (like an ad blocker) blocked the request.',
-		},
-		{
-			keyDisplayValue: 'Time',
-			valueDisplayValue:
-				selectedNetworkResource?.responseEnd &&
-				selectedNetworkResource?.startTime &&
-				formatTime(
-					selectedNetworkResource.responseEnd -
-						selectedNetworkResource.startTime,
-				),
-		},
-		{
-			keyDisplayValue: 'Type',
-			valueDisplayValue: getNetworkResourcesDisplayName(
-				selectedNetworkResource?.initiatorType || '',
-			),
-		},
-		...(selectedNetworkResource?.initiatorType === 'websocket'
-			? []
-			: [
-					{
-						keyDisplayValue: 'Size',
-						valueDisplayValue: selectedNetworkResource
-							?.requestResponsePairs?.response.size ? (
-							formatSize(
-								selectedNetworkResource.requestResponsePairs
-									.response.size,
-							)
-						) : selectedNetworkResource?.requestResponsePairs
-								?.response.status === 0 ? (
-							'-'
-						) : selectedNetworkResource?.requestResponsePairs
-								?.urlBlocked ||
-						  selectedNetworkResource?.transferSize == null ? (
-							'-'
-						) : selectedNetworkResource?.transferSize === 0 ? (
-							'Cached'
-						) : (
-							<>
-								{formatSize(
-									selectedNetworkResource.transferSize,
-								)}
-							</>
-						),
-					},
-			  ]),
-	]
-
-	const showRequestMetrics =
-		selectedNetworkResource &&
-		REQUEST_INITIATOR_TYPES.indexOf(
-			selectedNetworkResource.initiatorType as typeof REQUEST_INITIATOR_TYPES[number],
-		) > -1
-
-	if (selectedNetworkResource?.requestResponsePairs) {
-		const { request, response, urlBlocked } =
-			selectedNetworkResource.requestResponsePairs
-
-		if (urlBlocked) {
-			generalData.push({
-				keyDisplayValue: 'Recording Blocked',
-				valueDisplayValue: (
-					<>
-						<span>
-							The headers and body of this request was blocked.
-							The URL matched one a URL that is known to contain
-							secrets/passwords. You can{' '}
-							<a
-								href="https://docs.highlight.run/recording-network-requests-and-responses"
-								target="_blank"
-								rel="noreferrer"
-							>
-								learn more here.
-							</a>
-						</span>
-					</>
-				),
-			})
-		}
-
-		if (request.headers) {
-			const requestHeaderKeys = Object.keys(request.headers)
-
-			requestHeaderKeys.forEach((key) => {
-				requestHeadersData.push({
-					keyDisplayValue: key,
-					valueDisplayValue: request.headers[key],
-				})
-			})
-		}
-
-		if (request.body) {
-			try {
-				const parsedRequestBody = JSON.parse(request.body)
-
-				Object.keys(parsedRequestBody).forEach((key) => {
-					// `query` is a special for GraphQL requests.
-					// Check to see if the value for `query` is valid GraphQL, if so render it using a GraphQL formatter
-					if (key === 'query') {
-						const queryString = parsedRequestBody[key]
-
-						requestPayloadData.push({
-							keyDisplayValue: key,
-							valueDisplayValue: (
-								<CodeBlock
-									language="graphql"
-									text={queryString}
-									wrapLines
-									wrapLongLines
-									hideCopy
-									customStyle={{
-										fontSize: '10px',
-									}}
-								/>
-							),
-							data: queryString,
-						})
-					} else {
-						const renderType =
-							typeof parsedRequestBody[key] === 'object'
-								? 'json'
-								: 'string'
-						requestPayloadData.push({
-							keyDisplayValue: key,
-							valueDisplayValue:
-								renderType === 'string'
-									? parsedRequestBody[key]?.toString()
-									: JSON.parse(
-											JSON.stringify(
-												parsedRequestBody[key],
-											),
-									  ),
-						})
-					}
-				})
-			} catch {
-				requestPayloadData.push({
-					keyDisplayValue: 'body',
-					valueDisplayValue: JSON.stringify(request.body),
-				})
-			}
-		}
-
-		if (response.headers) {
-			Object.keys(response.headers).forEach((key) => {
-				responseHeadersData.push({
-					keyDisplayValue: key,
-					valueDisplayValue: response.headers[key]?.toString(),
-				})
-			})
-		}
-
-		if (response.size) {
-			responsePayloadData.push({
-				keyDisplayValue: 'Size',
-				valueDisplayValue: formatSize(response.size),
-			})
-		}
-
-		if (response.body) {
-			try {
-				let parsedResponseBody: { [key: string]: any } = {}
-				if (typeof response.body === 'object') {
-					parsedResponseBody = JSON.parse(
-						JSON.stringify(response.body),
-					)
-				} else {
-					parsedResponseBody = JSON.parse(response.body)
-				}
-				Object.keys(parsedResponseBody).forEach((key) => {
-					const renderType =
-						typeof parsedResponseBody[key] === 'object'
-							? 'json'
-							: 'string'
-
-					responsePayloadData.push({
-						keyDisplayValue: key,
-						valueDisplayValue:
-							renderType === 'string'
-								? parsedResponseBody[key]?.toString()
-								: parsedResponseBody[key],
-					})
-				})
-			} catch (e) {
-				responsePayloadData.push({
-					keyDisplayValue: '-',
-					valueDisplayValue: JSON.stringify(response.body),
-				})
-			}
-		}
-	}
-
-	const getErrorRows = (errorObj: ErrorObject): TableListItem[] => [
-		{
-			keyDisplayValue: 'Event',
-			valueDisplayValue: errorObj.event.find((e) => e),
-		},
-		{
-			keyDisplayValue: 'Stack Trace',
-			valueDisplayValue: errorObj.stack_trace,
-		},
-	]
-
-	const sections: [NetworkResourceMeta, TableListItem[]][] = [
-		[NetworkResourceMeta.General, generalData],
-	]
-
-	if (selectedNetworkResource?.errors?.length) {
-		selectedNetworkResource?.errors?.forEach((error, idx) => {
-			sections.push([
-				`Error ${idx + 1}` as NetworkResourceMeta,
-				getErrorRows(error),
-			])
-		})
-	}
-
-	if (
-		(selectedNetworkResource?.initiatorType === 'fetch' ||
-			selectedNetworkResource?.initiatorType === 'xmlhttprequest') &&
-		!selectedNetworkResource.requestResponsePairs?.urlBlocked
-	) {
-		Object.entries({
-			[NetworkResourceMeta.RequestHeaders]: requestHeadersData,
-			[NetworkResourceMeta.RequestPayload]: requestPayloadData,
-			[NetworkResourceMeta.ResponseHeaders]: responseHeadersData,
-			[NetworkResourceMeta.ResponsePayload]: responsePayloadData,
-		}).forEach(([key, value]) =>
-			sections.push([key as NetworkResourceMeta, value]),
-		)
-	}
-
-	return (
-		<Box>
-			{showRequestMetrics && (
-				<RequestMetrics resource={selectedNetworkResource} />
-			)}
-
-			{sections.map(([key, value]) => (
-				<Box key={key} p="20">
-					<Box
-						pb="12"
-						bb="secondary"
-						display="flex"
-						justifyContent="space-between"
-						alignItems="center"
-					>
-						<Text size="large" weight="bold" color="strong">
-							{key}
-						</Text>
-					</Box>
-
-					<Box
-						mt="12"
-						display="flex"
-						justifyContent="space-between"
-						alignItems="center"
-					>
-						<TableList
-							data={value}
-							noDataMessage={
-								networkRecordingEnabledForSession ? (
-									<NoRecordingMessage />
-								) : (
-									<NetworkRecordingEducationMessage />
-								)
-							}
-						/>
-					</Box>
-				</Box>
-			))}
-		</Box>
-	)
-}
-
-function NoRecordingMessage() {
-	return (
-		<Box width="full" display="flex" flexDirection="column" gap="12" pb="4">
-			<Text size="small" color="moderate" weight="medium">
-				Network recording is on, but no data was recorded.
-			</Text>
-		</Box>
-	)
-}
-
-function NetworkRecordingEducationMessage() {
-	return (
-		<Box width="full" display="flex" flexDirection="column" gap="12" pb="4">
-			<Text size="small" color="moderate" weight="medium">
-				<code>recordHeadersAndBody</code> is disabled. If you would like
-				to see XHR/Fetch headers and bodies you will need to enable{' '}
-				<code>recordHeadersAndBody</code>.
-			</Text>
-			<Text size="small" color="moderate" weight="medium">
-				You can learn more about this and about the security/privacy
-				implications{' '}
-				<a
-					target="_blank"
-					rel="noreferrer"
-					href="https://docs.highlight.run/recording-network-requests-and-responses"
-				>
-					here
-				</a>
-				.
-			</Text>
-		</Box>
-	)
-}
-
-// The amount of time before and after the request started/ended we want to show
-// logs for.
-const TIME_BUFFER = 200000
-
-const NetworkResourceLogs: React.FC<{
-	resource: NetworkResource
-	sessionStartTime: number
-}> = ({ resource, sessionStartTime }) => {
-	const { project_id } = useParams<{
-		project_id: string
-	}>()
-	const [query, setQuery] = useState(
-		stringifyLogsQuery([
-			{
-				key: 'trace_id',
-				operator: DEFAULT_LOGS_OPERATOR,
-				value: String(resource.id),
-				offsetStart: 0,
-			},
-		]),
-	)
-	const tableContainerRef = useRef<HTMLDivElement>(null)
-	const startDate = useMemo(
-		() => new Date(sessionStartTime + resource.startTime - TIME_BUFFER),
-		[sessionStartTime, resource.startTime],
-	)
-	const endDate = useMemo(
-		() => new Date(sessionStartTime + resource.responseEnd + TIME_BUFFER),
-		[resource.responseEnd, sessionStartTime],
-	)
-
-	const {
-		logEdges,
-		loading,
-		error,
-		loadingAfter,
-		fetchMoreForward,
-		fetchMoreBackward,
-		refetch,
-	} = useGetLogs({
-		query,
-		project_id,
-		logCursor: undefined,
-		startDate,
-		endDate,
-	})
-
-	const fetchMoreWhenScrolled = React.useCallback(
-		(containerRefElement?: HTMLDivElement | null) => {
-			if (containerRefElement) {
-				const { scrollHeight, scrollTop, clientHeight } =
-					containerRefElement
-
-				if (scrollHeight - scrollTop - clientHeight < 100) {
-					fetchMoreForward()
-				} else if (scrollTop === 0) {
-					fetchMoreBackward()
-				}
-			}
-		},
-		[fetchMoreForward, fetchMoreBackward],
-	)
-
-	return (
-		<>
-			<Box
-				padding="8"
-				flex="stretch"
-				justifyContent="stretch"
-				display="flex"
-				overflow="hidden"
-				maxHeight="full"
-			>
-				<Box
-					borderRadius="6"
-					flexDirection="column"
-					display="flex"
-					flexGrow={1}
-					border="dividerWeak"
-					shadow="medium"
-				>
-					<SearchForm
-						initialQuery={query}
-						onFormSubmit={(value) => setQuery(value)}
-						startDate={startDate}
-						endDate={endDate}
-						onDatesChange={() => null}
-						presets={[]}
-						minDate={new Date(sessionStartTime)}
-						timeMode="permalink"
-						disableSearch
-						addLinkToViewInLogViewer
-						hideDatePicker
-						hideCreateAlert
-					/>
-					<Box
-						height="screen"
-						pt="4"
-						px="12"
-						pb="12"
-						overflowY="auto"
-						onScroll={(e) =>
-							fetchMoreWhenScrolled(e.target as HTMLDivElement)
-						}
-						ref={tableContainerRef}
-					>
-						{logEdges.length === 0 ? (
-							<NoLogsFound />
-						) : (
-							<LogsTable
-								logEdges={logEdges}
-								loading={loading}
-								error={error}
-								refetch={refetch}
-								loadingAfter={loadingAfter}
-								query={query}
-								tableContainerRef={tableContainerRef}
-								selectedCursor={undefined}
-							/>
-						)}
-					</Box>
-				</Box>
-			</Box>
-		</>
-	)
-}
-
-const ErrorsData: React.FC<{ resource: NetworkResource; hide: () => void }> = ({
-	resource,
-	hide,
-}) => {
-	const { projectId } = useProjectId()
-	const { errors: sessionErrors } = useReplayerContext()
-	const requestId = getHighlightRequestId(resource)
-	const errors = sessionErrors.filter((e) => e.request_id === requestId)
-	const errorGroupSecureIds = errors.map((e) => e.error_group_secure_id)
-	const { data, loading } = useGetErrorGroupsOpenSearchQuery({
-		variables: {
-			query: `{
-				"bool": {
-					"must": {
-						"terms": {
-							"secure_id.keyword": [${errorGroupSecureIds.map((id) => `"${id}"`).join(',')}]
-						}
-					}
-				}
-			}`.replace(/\s+/g, ''),
-			project_id: projectId,
-			count: errorGroupSecureIds.length,
-		},
-		skip: errors.length === 0,
-	})
-
-	const { setActiveError, setRightPanelView } = usePlayerUIContext()
-	const { setShowRightPanel } = usePlayerConfiguration()
-
-	return (
-		<>
-			{data?.error_groups_opensearch.error_groups?.length ? (
-				data?.error_groups_opensearch.error_groups.map(
-					(errorGroup, idx) => (
-						<Box
-							py="8"
-							px="12"
-							flex="stretch"
-							justifyContent="stretch"
-							display="flex"
-							key={idx}
-						>
-							<ErrorFeedCard
-								errorGroup={errorGroup}
-								onClick={() => {
-									const error = errors.find(
-										(e) =>
-											e.error_group_secure_id ===
-											errorGroup.secure_id,
-									)
-									setActiveError(error)
-									setShowRightPanel(true)
-									setRightPanelView(RightPanelView.Error)
-									hide()
-								}}
-							/>
-						</Box>
-					),
-				)
-			) : loading ? (
-				<LoadingBox />
-			) : (
-				<FullScreenContainer>
-					<Box style={{ maxWidth: 300 }}>
-						<Callout title="No errors">
-							<Box mb="6">
-								<Text>
-									There are no errors associated with this
-									network request.
-								</Text>
-							</Box>
-						</Callout>
-					</Box>
-				</FullScreenContainer>
-			)}
-		</>
-	)
-}
-
-const NoLogsFound = () => {
-	return (
-		<FullScreenContainer>
-			<Box style={{ maxWidth: 300 }}>
-				<Callout title="No associated logs found">
-					<Box
-						display="flex"
-						flexDirection="column"
-						gap="16"
-						alignItems="flex-start"
-					>
-						<Text color="moderate">
-							To match backend logs to network request make sure
-							to enable "full stack mapping."
-						</Text>
-
-						<LinkButton
-							trackingId="logs-empty-state_specification-docs"
-							kind="secondary"
-							to="https://www.highlight.io/docs/getting-started/frontend-backend-mapping"
-							target="_blank"
-						>
-							Learn more
-						</LinkButton>
-					</Box>
-				</Callout>
-			</Box>
-		</FullScreenContainer>
 	)
 }
